Move patched files restoration too
