+++ src/Microsoft.DotNet.Darc/src/DarcLib/VirtualMonoRepo/VmrPatchHandler.cs
@@ -2,11 +2,14 @@
 // The .NET Foundation licenses this file to you under the MIT license.
 // See the LICENSE file in the project root for more information.
 
+using System;
 using System.Collections.Generic;
 using System.IO;
 using System.Linq;
+using System.Text.RegularExpressions;
 using System.Threading;
 using System.Threading.Tasks;
+using LibGit2Sharp;
 using Microsoft.DotNet.Darc.Models.VirtualMonoRepo;
 using Microsoft.DotNet.DarcLib.Helpers;
 using Microsoft.Extensions.Logging;
@@ -14,25 +17,23 @@ using Microsoft.Extensions.Logging;
 #nullable enable
 namespace Microsoft.DotNet.DarcLib.VirtualMonoRepo;
 
-public interface IVmrPatchHandler
-{
-    Task ApplyPatch(SourceMapping mapping, string patchPath, CancellationToken cancellationToken);
-    
-    Task CreatePatch(
-        SourceMapping mapping,
-        string repoPath,
-        string sha1,
-        string sha2,
-        string destPath,
-        CancellationToken cancellationToken);
-
-    Task ApplyVmrPatches(SourceMapping mapping, CancellationToken cancellationToken);
-}
-
+/// <summary>
+/// This class handles git patches during VMR synchronization.
+/// It can create/apply diffs and handle VMR patches which are additional patches stored in the VMR
+/// that are applied on top of individual repositories.
+/// </summary>
 public class VmrPatchHandler : IVmrPatchHandler
 {
     private const string KeepAttribute = "vmr-preserve";
     private const string IgnoreAttribute = "vmr-ignore";
+
+    /// <summary>
+    /// Matches output of `git apply --numstat` which lists files contained in a patch file.
+    /// Example output:
+    /// 0       14      /s/vmr/src/roslyn-analyzers/eng/Versions.props
+    /// -       -       /s/vmr/src/roslyn-analyzers/some-binary.dll
+    /// </summary>
+    private static readonly Regex GitPatchSummaryLine = new(@"^[\-0-9]+\s+[\-0-9]+\s+(?<file>[^\s]+)$", RegexOptions.Compiled);
 
     private readonly IVmrDependencyTracker _vmrInfo;
     private readonly IProcessManager _processManager;
@@ -172,6 +173,59 @@ public class VmrPatchHandler : IVmrPatchHandler
     }
 
     /// <summary>
+    /// For all files for which we have patches in VMR, restore their original version from the repo.
+    /// This is because VMR contains already patched versions of these files and new updates from the repo wouldn't apply.
+    /// </summary>
+    /// <param name="mapping">Mapping</param>
+    /// <param name="clonePath">Path were the individual repo was cloned to</param>
+    /// <param name="originalRevision">Revision from which we were updating</param>
+    public async Task RestorePatchedFilesFromRepo(
+        SourceMapping mapping,
+        string clonePath,
+        string originalRevision,
+        CancellationToken cancellationToken)
+    {
+        if (!mapping.VmrPatches.Any())
+        {
+            return;
+        }
+
+        _logger.LogInformation("Restoring files with patches for {mappingName}..", mapping.Name);
+
+        var localRepo = new LocalGitClient(_processManager.GitExecutable, _logger);
+        localRepo.Checkout(clonePath, originalRevision);
+
+        var repoSourcesPath = _vmrInfo.GetRepoSourcesPath(mapping);
+
+        foreach (var patch in mapping.VmrPatches)
+        {
+            _logger.LogDebug("Processing VMR patch `{patch}`..", patch);
+
+            foreach (var patchedFile in await GetFilesInPatch(clonePath, patch, cancellationToken))
+            {
+                // git always works with forward slashes (even on Windows)
+                string relativePath = Path.DirectorySeparatorChar != '/'
+                    ? patchedFile.Replace('/', Path.DirectorySeparatorChar)
+                    : patchedFile;
+
+                var originalFile = Path.Combine(clonePath, relativePath);
+                var destination = Path.Combine(repoSourcesPath, relativePath);
+
+                _logger.LogDebug("Restoring file `{originalFile}` to `{destination}`..", originalFile, destination);
+
+                // Copy old revision to VMR
+                File.Copy(originalFile, destination, overwrite: true);
+            }
+        }
+
+        // Stage the restored files (all future patches are applied to index directly)
+        using var repository = new Repository(_vmrInfo.VmrPath);
+        Commands.Stage(repository, repoSourcesPath);
+
+        _logger.LogDebug("Files from VMR patches for {mappingName} restored", mapping.Name);
+    }
+
+    /// <summary>
     /// Applies VMR patches onto files of given mapping's subrepository.
     /// These files are stored in the VMR and applied on top of the individual repos.
     /// </summary>
@@ -192,5 +246,29 @@ public class VmrPatchHandler : IVmrPatchHandler
             _logger.LogDebug("Applying {patch}..", patchFile);
             await ApplyPatch(mapping, patchFile, cancellationToken);
         }
+    }
+
+    /// <summary>
+    /// Resolves a list of all files that are part of a given patch diff.
+    /// </summary>
+    /// <param name="repoPath">Path (to the repo) the patch applies onto</param>
+    /// <param name="patchPath">Path to the patch file</param>
+    /// <returns>List of all files (paths relative to repo's root) that are part of a given patch diff</returns>
+    private async Task<IReadOnlyCollection<string>> GetFilesInPatch(string repoPath, string patchPath, CancellationToken cancellationToken)
+    {
+        var result = await _processManager.ExecuteGit(repoPath, new[] { "apply", "--numstat", patchPath }, cancellationToken);
+        result.ThrowIfFailed($"Failed to enumerate files from a patch at `{patchPath}`");
+
+        var files = new List<string>();
+        foreach (var line in result.StandardOutput.Split(Environment.NewLine))
+        {
+            var match = GitPatchSummaryLine.Match(line);
+            if (match.Success)
+            {
+                files.Add(match.Groups["file"].Value);
+            }
+        }
+
+        return files;
     }
 }
